# PERF: isinstance ABCIndexClass and ABCExtensionArray